Add simple test

+++ system.hpp
@@ -37,9 +37,11 @@ namespace ecs
             auto filtered = m_ecsManager->getEntities();
             for (auto it = filtered.begin(); it != filtered.end();) {
                 auto components = it->second->getComponents();
-                if (std::any_of(m_componentTypes.begin(), m_componentTypes.end(),
+                if (std::any_of(m_componentTypes.begin(),
+                                m_componentTypes.end(),
                                 [&components](size_t t) {
-                                    return components.find(t) == components.end();
+                                    return components.find(t) ==
+                                           components.end();
                                 }))
                     it = filtered.erase(it);
                 else
@@ -57,8 +59,9 @@ namespace ecs
         template<typename... ComponentTypes>
         auto getEntitiesByTags() const
         {
-            static_assert(types::IsBaseOfRec<Component, types::TypeList<ComponentTypes...>>::value,
-                          "Template parameter class must be child of Component");
+            static_assert(
+                    types::IsBaseOfRec<Component, types::TypeList<ComponentTypes...>>::value,
+                    "Template parameter class must be child of Component");
             using ComponentList = types::TypeList<ComponentTypes...>;
             static_assert(types::Length<ComponentList>::value >= 2,
                           "Length of ComponentTypes must be greeter than 2");
@@ -89,7 +92,7 @@ namespace ecs
         template<class ComponentType>
         auto getEntitiesByTag() const
         {
-            static_assert(std::is_base_of_v<Component, ComponentType>,
+            static_assert(std::is_base_of_v < Component, ComponentType > ,
                           "ComponentType class must be child of Component");
             auto filtered = m_ecsManager->getEntities();
             for (auto it = filtered.begin(); it != filtered.end();) {
@@ -104,7 +107,7 @@ namespace ecs
 
     private:
         // Contains id's of each component type system can handle
-        std::set<size_t> m_componentTypes;
+        std::set<int> m_componentTypes;
     };
 };
